Combining work on semantic-ui.
@ambermanry your files have been merged into the structure. Instead of
the template applying directly to the body it is now on the lunchroom
template.

the eventList, and newEvent are now found in feature directories under
event/features

the styles have been added into styles.css

the events collections declaration has been moved to collections.js

the helpers for ‘events’ that was on body has been moved to lunchroom.